Enhanced Vehicle and Contract models with new attributes; improved key logic in methods and implemented GET endpoint in ContractsController

+++ backend/Controllers/VehiclesController.cs
@@ -62,6 +62,11 @@ namespace JAPChallenge.Controllers
                 return NotFound();
             }
 
+            if (existingVehicle.Status == "Alugado")
+            {
+                return Conflict(new { message = "The vehicle is currently rented and cannot be deleted." });
+            }
+
             _context.Vehicles.Remove(existingVehicle);
             await _context.SaveChangesAsync();
